+++ apps/fabric-website/just-task.js
@@ -1,0 +1,9 @@
+const preset = require('@uifabric/build/just-task');
+const { task, option } = preset.just;
+const createFlightConfigTaskFactory = require('./scripts/createFlightConfig');
+
+module.exports = function() {
+  preset();
+  option('baseCDNUrl', { default: './dist' });
+  task('create-flight-config', createFlightConfigTaskFactory());
+};