+++ apps/fabric-website/src/components/App/AppState.tsx
@@ -1,13 +1,17 @@
 import * as React from 'react';
 
 // Props
-import { INavPage } from '../Nav/Nav.types';
 import { ComponentPage } from '../ComponentPage/ComponentPage';
 import { PageHeader } from '../PageHeader/PageHeader';
+import { ApiReferencesTableSet } from '@uifabric/example-app-base';
+import { IReferencesList } from '@uifabric/fabric-website-resources/lib/AppDefinition';
+const pageStyles: any = require('../../pages/PageStyles.module.scss');
+
+export interface ILegacyNavPage {}
 
 export interface IAppState {
   appTitle: string;
-  pages: INavPage[];
+  pages: ILegacyNavPage[];
 }
 
 // Giving the loading component a height so that the left nav loads in full screen and there is less flashing as the component page loads.
@@ -28,6 +32,27 @@ const StylesLoadingComponent = (props: any): JSX.Element => {
     </div>
   );
 };
+
+function loadReferences(): ILegacyNavPage[] {
+  const requireContext = require.context('@uifabric/api-docs/lib/pages/references', false, /\w+\.page\.json$/);
+
+  return requireContext.keys().map(pagePath => {
+    const pageName = pagePath.match(/(\w+)\.page\.json/)![1];
+    return {
+      title: pageName,
+      url: '#/components/references/' + pageName.toLowerCase(),
+      isFilterable: true,
+      component: () => (
+        <div className={pageStyles.basePage}>
+          <ComponentPage>
+            <PageHeader pageTitle={pageName} backgroundColor="#038387" />
+            <ApiReferencesTableSet jsonDocs={require('@uifabric/api-docs/lib/pages/references/' + pageName + '.page.json')} />
+          </ComponentPage>
+        </div>
+      )
+    };
+  });
+}
 
 export const AppState: IAppState = {
   appTitle: 'Office UI Fabric',
@@ -157,7 +182,7 @@ export const AppState: IAppState = {
             },
             {
               title: 'ComboBox',
-              url: '#/components/ComboBox',
+              url: '#/components/combobox',
               isFilterable: true,
               component: () => <LoadingComponent title="ComboBox" />,
               getComponent: cb =>
@@ -329,7 +354,7 @@ export const AppState: IAppState = {
             },
             {
               title: 'Calendar',
-              url: '#/components/Calendar',
+              url: '#/components/calendar',
               isFilterable: true,
               component: () => <LoadingComponent title="Calendar" />,
               getComponent: cb =>
@@ -930,6 +955,13 @@ export const AppState: IAppState = {
                 require.ensure([], require => cb(require<any>('../../pages/Components/KeytipsComponentPage').KeytipsComponentPage))
             }
           ]
+        },
+        {
+          title: 'References',
+          url: '#/components/references',
+          className: 'componentsPage',
+          isCategory: true,
+          pages: loadReferences()
         }
       ]
     },