+++ apps/fabric-website/src/pages/Controls/FocusTrapZonePage/FocusTrapZonePage.tsx
@@ -1,0 +1,7 @@
+import * as React from 'react';
+import { ControlsAreaPage, IControlsPageProps } from '../ControlsAreaPage';
+import { FocusTrapZonePageProps } from './FocusTrapZonePage.doc';
+
+export const FocusTrapZonePage: React.StatelessComponent<IControlsPageProps> = props => {
+  return <ControlsAreaPage {...props} {...FocusTrapZonePageProps[props.platform]} />;
+};