+++ packages/office-ui-fabric-react/src/components/ComboBox/ComboBox.doc.tsx
@@ -58,7 +58,6 @@ export const ComboBoxPageProps: IDocPageProps = {
       codepenJS: ComboBoxCustomStyledExampleCodepen
     }
   ],
-  propertiesTablesSources: [require<string>('!raw-loader!office-ui-fabric-react/src/components/ComboBox/ComboBox.types.ts')],
   overview: require<string>('!raw-loader!office-ui-fabric-react/src/components/ComboBox/docs/ComboBoxOverview.md'),
   bestPractices: '',
   dos: require<string>('!raw-loader!office-ui-fabric-react/src/components/ComboBox/docs/ComboBoxDos.md'),