+++ packages/office-ui-fabric-react/src/components/DocumentCard/DocumentCardTitle.base.tsx
@@ -21,6 +21,9 @@ const TRUNCATION_MAX_LENGTH_SECONDARY = 130 - TRUNCATION_SEPARATOR.length;
 const TRUNCATION_FIRST_PIECE_LONGER_BY = 10;
 const TRUNCATION_VERTICAL_OVERFLOW_THRESHOLD = 5;
 
+/**
+ * {@docCategory DocumentCard}
+ */
 export class DocumentCardTitleBase extends BaseComponent<IDocumentCardTitleProps, IDocumentCardTitleState> {
   private _titleElement = React.createRef<HTMLDivElement>();
   private _scrollTimerId: number;