+++ packages/office-ui-fabric-react/src/components/MessageBar/MessageBar.doc.tsx
@@ -28,7 +28,6 @@ export const MessageBarPageProps: IDocPageProps = {
       codepenJS: MessageBarStyledExampleCodepen
     }
   ],
-  propertiesTablesSources: [require<string>('!raw-loader!office-ui-fabric-react/src/components/MessageBar/MessageBar.types.ts')],
   overview: require<string>('!raw-loader!office-ui-fabric-react/src/components/MessageBar/docs/MessageBarOverview.md'),
   bestPractices: '',
   dos: require<string>('!raw-loader!office-ui-fabric-react/src/components/MessageBar/docs/MessageBarDos.md'),