+++ packages/office-ui-fabric-react/src/components/Modal/Modal.styles.ts
@@ -23,7 +23,8 @@ export const getStyles = (props: IModalStyleProps): IModalStyles => {
     theme,
     topOffsetFixed,
     isModeless,
-    layerClassName
+    layerClassName,
+    isDefaultDragHandle
   } = props;
   const { palette, effects } = theme;
 
@@ -74,6 +75,9 @@ export const getStyles = (props: IModalStyleProps): IModalStyles => {
         hasBeenOpened && {
           top: modalRectangleTop
         },
+      isDefaultDragHandle && {
+        cursor: 'move'
+      },
       containerClassName
     ],
     scrollableContent: [
@@ -98,6 +102,17 @@ export const getStyles = (props: IModalStyleProps): IModalStyles => {
         width: 'unset',
         height: 'unset'
       }
-    ]
+    ],
+    keyboardMoveIconContainer: {
+      position: 'absolute',
+      display: 'flex',
+      justifyContent: 'center',
+      width: '100%',
+      padding: '3px 0px'
+    },
+    keyboardMoveIcon: {
+      fontSize: '24px',
+      width: '24px'
+    }
   };
 };