+++ packages/office-ui-fabric-react/src/components/pickers/BasePicker.types.ts
@@ -1,6 +1,5 @@
 import * as React from 'react';
 import { IPickerItemProps } from './PickerItem.types';
-import { IPersonaProps } from '../Persona/Persona.types';
 import { IRefObject, IStyleFunctionOrObject } from '../../Utilities';
 import { ISuggestionModel, ISuggestionsProps } from './Suggestions/Suggestions.types';
 import { BaseAutoFill } from './AutoFill/BaseAutoFill';
@@ -116,9 +115,9 @@ export interface IBasePickerProps<T> extends React.Props<any> {
   inputProps?: IInputProps;
 
   /**
-   * A callback for when a persona is removed from the suggestion list
+   * A callback for when an item is removed from the suggestion list
    */
-  onRemoveSuggestion?: (item: IPersonaProps) => void;
+  onRemoveSuggestion?: (item: T) => void;
 
   /**
    * A function used to validate if raw text entered into the well can be added into the selected items list